+++ CHANGELOG.md
@@ -17,6 +17,7 @@ IMPROVEMENTS:
 
 BUG FIXES:
 
+- Fix pvtz 'Zone.VpcExists' error [GH-690]
 - Fix multi-k8s testcase failed error [GH-683]
 - Fix pvtz attchment Zone.NotExists error [GH-682]
 - Fix deleting ram role error [GH-674]